Feature gate native in stf runner
