+++ full-node/sov-stf-runner/Cargo.toml
@@ -23,11 +23,11 @@ tracing = { workspace = true }
 futures = "0.3"
 tracing-subscriber = "0.3.17"
 
-sov-db = { path = "../db/sov-db" }
+sov-db = { path = "../db/sov-db", optional = true }
 sov-rollup-interface = { path = "../../rollup-interface", version = "0.1" }
-sov-state = { path = "../../module-system/sov-state", version = "0.1", features = ["native"] }
-sov-modules-api = { path = "../../module-system/sov-modules-api", version = "0.1", features = ["native"] }
-celestia = { path = "../../adapters/celestia", features = ["native"] }
+sov-state = { path = "../../module-system/sov-state", version = "0.1" }
+sov-modules-api = { path = "../../module-system/sov-modules-api", version = "0.1" }
+celestia = { path = "../../adapters/celestia" }
 
 [dev-dependencies]
 tempfile = { workspace = true }
@@ -39,12 +39,13 @@ sov-modules-stf-template = { path = "../../module-system/sov-modules-stf-templat
 sov-value-setter = { path = "../../module-system/module-implementations/examples/sov-value-setter", features = ["native"] }
 sov-accounts = { path = "../../module-system/module-implementations/sov-accounts", features = ["native"] }
 
-#[features]
-#default = []
-#native = [
-#    "sov-sequencer-registry/native",
-#    "sov-bank/native",
-#    "sov-value-setter/native",
-#    "sov-accounts/native",
-#    "jupiter/native",
-#]
+[features]
+default = []
+native = [
+	"sov-db",
+	"celestia/native",
+    "sov-sequencer-registry/native",
+    "sov-bank/native",
+    "sov-state/native",
+    "sov-modules-api/native",
+]
